fix(diagnostics): Replace download link for Spring Boot apps
Due to an error when using Spring Boot apps together with
Hawtio version <2.15 the download link cannot be used the same way as
other mBean-enabled contexts.

Here, a workaround is implemented to save the Flight Recorder recording
on the host machine by clicking the download button. The recording will
be saved as: <recording id>.jfr.
The best approach would be to fix the error in the downloadRecording
implementation in the Hawtio backend, but that is for the time being
substantially more difficult.

Fixes #2724 in hawtio issue tracker https://github.com/hawtio/hawtio/issues/2724